Create manage_courses app  succesfully
