+++ Backend/Learnify/urls.py
@@ -16,7 +16,7 @@ Including another URLconf
 """
 from django.contrib import admin
 from django.urls import path, include
-from learnify_app.urls import urlpatterns as learnify_app_urls #Importing the urls from the learnify_app to be used in the
+from auth_users.urls import urlpatterns as auth_users_urls #Importing the urls from the auth_users to be used in the
 
 
 
@@ -25,6 +25,6 @@ from learnify_app.urls import urlpatterns as learnify_app_urls #Importing the ur
 urlpatterns = [
     path('admin/', admin.site.urls),
     path('api/auth/', include('dj_rest_auth.urls')), #Endpoint provided by dj_rest_aut
-    path('api/social/login/', include(learnify_app_urls)), #Endpoint provided by dj_rest_auth
+    path('api/social/login/', include(auth_users_urls)), #Endpoint provided by dj_rest_auth
 ]
