+++ Backend/learnify_app/urls.py
@@ -1,6 +1,1 @@
 from django.urls import path
-from .views import GoogleLoginView  
-
-urlpatterns = [
-    path('google/', GoogleLoginView.as_view(), name='google_login'),
-]
