+++ Backend/manage_courses/models.py
@@ -1,3 +1,41 @@
+import uuid
 from django.db import models
+from auth_users.models import User
 
-# Create your models here.
+class Course(models.Model):
+    instructor = models.ForeignKey(User, on_delete=models.CASCADE, related_name="courses")
+    title = models.CharField(max_length=255)
+    description = models.TextField()
+    price = models.DecimalField(max_digits=15, decimal_places=2)
+    category = models.CharField(max_length=50)
+    status = models.CharField(max_length=15, choices=[("draft", "Draft"), ("published", "Published"), ("archived", "Archived")])
+    created_at = models.DateTimeField(auto_now_add=True)
+    
+    class Meta:
+        verbose_name = "Course"
+
+    def __str__(self):
+        return f"{self.title} - {self.instructor}"
+
+class Module(models.Model):
+    id = models.UUIDField(default=uuid.uuid4, primary_key=True, editable=False)
+    course = models.ForeignKey(Course, on_delete=models.CASCADE, related_name="modules")
+    title = models.CharField(max_length=255)
+    position = models.IntegerField()
+    
+    class Meta:
+        verbose_name = "Module"
+    
+    def __str__(self):
+        return f"{self.title} - {self.course}"
+
+class Lesson(models.Model):
+    id = models.UUIDField(default=uuid.uuid4, primary_key=True, editable=False)
+    module = models.ForeignKey(Module, on_delete=models.CASCADE, related_name="lessons")
+    title = models.CharField(max_length=255)
+    content= models.TextField()
+    video_url = models.URLField(blank=True, null=True)
+    position = models.IntegerField()
+
+    def __str__(self):
+        return f"{self.title} - {self.module}"
